WIP: Test for TLS websockets
This is meant to be a test for #18977, but encountering a different issue on the potentially fixed branch, at the moment.

A bit stuck as the client here gets an 'InvalidData: invalid data' error right away.
Opening a WebSocket from the browser to the server running in the test is fine.
And the test passes if the cert/key are commented out and wss -> ws.